requirements.txt
appdirs>=1.4.3
boilerpipe3>=1.1
charade>=1.0.3
JPype1-py3>=0.5.5.2
packaging>=16.8
pypandoc>=1.3.3
pyparsing>=2.2.0
requests>=2.13.0
six>=1.10.0
